Add files via upload
This is the main.c file for the bobbel sort algorithem:
it contains one sorting algorithem for numbers and one for letters:
It is in ascending order for number and also for letters, letters sorted though their ASCII value. It is therefore also case sensitive and sorts uppercase letters first.